+++ src/SpectraProcessing.Domain/MathModeling/PeakModeling.cs
@@ -4,24 +4,24 @@ namespace SpectraProcessing.Domain.MathModeling;
 
 public static class PeakModeling
 {
-    public static double GetPeakValueAt(this IReadOnlyPeakData peak, double x)
+    public static float GetPeakValueAt(this IReadOnlyPeakData peak, float x)
     {
         return peak.GaussianContribution * Gaussian(x, peak)
             + (1 - peak.GaussianContribution) * Lorentzian(x, peak);
 
-        static double Gaussian(double x, IReadOnlyPeakData estimate)
+        static float Gaussian(float x, IReadOnlyPeakData estimate)
         {
             // constant = -4 * Math.Log(2)
-            const double constant = -2.7725887222397811;
+            const float constant = -2.7725887222397811f;
 
             var b = x - estimate.Center;
 
             var c = estimate.HalfWidth * estimate.HalfWidth;
 
-            return estimate.Amplitude * Math.Exp(constant * b * b / c);
+            return estimate.Amplitude * (float) Math.Exp(constant * b * b / c);
         }
 
-        static double Lorentzian(double x, IReadOnlyPeakData estimate)
+        static float Lorentzian(float x, IReadOnlyPeakData estimate)
         {
             var a = 2 * (x - estimate.Center) / estimate.HalfWidth;
 
@@ -29,63 +29,63 @@ public static class PeakModeling
         }
     }
 
-    public static double GetPeaksValueAt(this IEnumerable<IReadOnlyPeakData> peaks, double x)
+    public static float GetPeaksValueAt(this IEnumerable<IReadOnlyPeakData> peaks, float x)
     {
         return peaks.Sum(p => p.GetPeakValueAt(x));
     }
 
-    public static double GetPeakArea(this IReadOnlyPeakData peak)
+    public static float GetPeakArea(this IReadOnlyPeakData peak)
     {
         return peak.GaussianContribution * GaussianSquare(peak)
             + (1 - peak.GaussianContribution) * LorentzianSquare(peak);
 
-        static double GaussianSquare(IReadOnlyPeakData peak)
+        static float GaussianSquare(IReadOnlyPeakData peak)
         {
             // constant = 2 * Math.Sqrt(Math.PI / Math.Log(2));
-            const double constant = 4.2578680777249049;
+            const float constant = 4.2578680777249049f;
 
             return constant * peak.Amplitude * peak.HalfWidth;
         }
 
-        static double LorentzianSquare(IReadOnlyPeakData peak)
+        static float LorentzianSquare(IReadOnlyPeakData peak)
         {
-            const double constant = Math.PI / 2;
+            const float constant = (float) Math.PI / 2;
 
             return constant * peak.Amplitude * peak.HalfWidth;
         }
     }
 
-    public static double GetPeakEffectiveRadius(this IReadOnlyPeakData peak, double areaPercentage)
+    public static float GetPeakEffectiveRadius(this IReadOnlyPeakData peak, float areaPercentage)
     {
         var gaussianWeight = GaussianAreaWeight(peak.GaussianContribution);
 
         return gaussianWeight * GaussianSquareRadius(peak, areaPercentage) +
             (1 - gaussianWeight) * LorentzianSquareRadius(peak, areaPercentage);
 
-        static double GaussianAreaWeight(double gaussianContribution)
+        static float GaussianAreaWeight(float gaussianContribution)
         {
             //gaussSquareDividedLorentzSquare = Math.Sqrt(Math.PI / Math.Log(2)) / Math.PI;
-            const double gaussSquareDividedLorentzSquare = 0.67766075160310502;
+            const float gaussSquareDividedLorentzSquare = 0.67766075160310502f;
             var gaussianArea = gaussianContribution * gaussSquareDividedLorentzSquare;
             return gaussianArea / (gaussianArea + (1 - gaussianContribution));
         }
 
-        static double GaussianSquareRadius(IReadOnlyPeakData peak, double areaPercentage)
+        static float GaussianSquareRadius(IReadOnlyPeakData peak, float areaPercentage)
         {
             // constant = 2 * Math.Sqrt(Math.Log(2));
-            const double constant = 1.6651092223153954;
+            const float constant = 1.6651092223153954f;
             return ErfInv(areaPercentage) * peak.HalfWidth / constant;
         }
 
-        static double LorentzianSquareRadius(IReadOnlyPeakData peak, double areaPercentage)
+        static float LorentzianSquareRadius(IReadOnlyPeakData peak, float areaPercentage)
         {
-            const double constant = Math.PI / 2;
+            const float constant = (float) Math.PI / 2;
 
-            return peak.HalfWidth * Math.Tan(constant * areaPercentage);
+            return peak.HalfWidth * (float) Math.Tan(constant * areaPercentage);
         }
     }
 
-    private static double Erf(double x)
+    private static float Erf(float x)
     {
         x = Math.Abs(x);
 
@@ -94,10 +94,10 @@ public static class PeakModeling
         var y = 1.0 - (((((1.061405429 * t - 1.453152027) * t) + 1.421413741) *
             t - 0.284496736) * t + 0.254829592) * t * Math.Exp(-x * x);
 
-        return Math.Sign(x) * y;
+        return Math.Sign(x) * (float) y;
     }
 
-    private static double ErfInv(double x)
+    private static float ErfInv(float x)
     {
         if (x < -1 || x > 1)
         {
@@ -109,11 +109,11 @@ public static class PeakModeling
             return 0;
         }
 
-        const double a = 0.147;
+        const float a = 0.147f;
         var ln = Math.Log(1 - x * x);
         var part1 = (2 / (Math.PI * a)) + (ln / 2);
         var part2 = ln / a;
 
-        return Math.Sign(x) * Math.Sqrt(Math.Sqrt(part1 * part1 - part2) - part1);
+        return Math.Sign(x) * (float) Math.Sqrt(Math.Sqrt(part1 * part1 - part2) - part1);
     }
 }
