+++ src/SpectraProcessing.Domain/Models/MathModeling/OptimizationSettings.cs
@@ -5,14 +5,14 @@ public sealed record OptimizationSettings
     public static readonly OptimizationSettings Default = new()
     {
         MaxIterationsCount = 10000,
-        MaxAcceptableRelativeDeviation = 0.001,
-        InitialShift = 0.01,
+        MaxAcceptableRelativeDeviation = 0.001f,
+        InitialShift = 0.01f,
         Coefficients = new IterationCoefficients
         {
             Reflection = 1,
             Expansion = 2,
-            Contraction = 0.5,
-            Shrink = 0.5,
+            Contraction = 0.5f,
+            Shrink = 0.5f,
         },
         Criteria = new CompletionСriteria
         {
@@ -23,9 +23,9 @@ public sealed record OptimizationSettings
 
     public required int MaxIterationsCount { get; init; }
 
-    public required double MaxAcceptableRelativeDeviation { get; init; }
+    public required float MaxAcceptableRelativeDeviation { get; init; }
 
-    public required double InitialShift { get; init; }
+    public required float InitialShift { get; init; }
 
     public required IterationCoefficients Coefficients { get; init; }
 
@@ -33,15 +33,15 @@ public sealed record OptimizationSettings
 
     public sealed record IterationCoefficients
     {
-        public required double Reflection { get; init; }
-        public required double Expansion { get; init; }
-        public required double Contraction { get; init; }
-        public required double Shrink { get; init; }
+        public required float Reflection { get; init; }
+        public required float Expansion { get; init; }
+        public required float Contraction { get; init; }
+        public required float Shrink { get; init; }
     }
 
     public sealed record CompletionСriteria
     {
-        public double? AbsoluteValue { get; init; }
+        public float? AbsoluteValue { get; init; }
 
         public int? MaxConsecutiveShrinks { get; init; }
     }
